src/test/java/com/ibc/LeaseServiceTest.java
package com.ibc;

import com.ibc.datautil.LeaseTestDataUtil;
import com.ibc.entity.Lease;
import com.ibc.entity.Car;
import com.ibc.exception.LeaseNotFoundException;
import com.ibc.repository.LeaseRepository;
import com.ibc.service.LeaseService;
import com.ibc.repository.CarRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;
import org.mockito.junit.jupiter.MockitoExtension;

import java.util.Optional;
import java.time.LocalDate;
import java.util.List;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.ArgumentMatchers.anyLong;
import static org.mockito.Mockito.*;

@ExtendWith(MockitoExtension.class)
public class LeaseServiceTest {

    @InjectMocks
    private LeaseService leaseService;

    @Mock
    private LeaseRepository leaseRepository;

    @Mock
    private CarRepository carRepository;

    @BeforeEach
    void setUp() {
        MockitoAnnotations.openMocks(this);
    }

    @Test
    void testCreateLease() {
        // Given
        Lease lease = LeaseTestDataUtil.getLeaseTestData().get(0);
        lease.setCar(new Car()); // Ensure car is not null
        lease.getCar().setPricePerDay(100.0);
        
        lease.setLeaseStartDate(LocalDate.now());
        lease.setLeaseEndDate(LocalDate.now().plusDays(5)); // Assuming 5 days of lease

        when(leaseRepository.save(any(Lease.class))).thenReturn(lease);

        // When
        Lease createdLease = leaseService.createLease(lease);

        // Then
        assertNotNull(createdLease);
        //assertEquals(600.0, createdLease.getTotalCost()); // Updated expected value
        verify(leaseRepository, times(1)).save(lease);
    }


    @Test
    void testReturnCar() {
        // Given
        Lease lease = LeaseTestDataUtil.getLeaseTestData().get(0);
        lease.setActive(true);
        when(leaseRepository.findById(anyLong())).thenReturn(Optional.of(lease));
        when(leaseRepository.save(any(Lease.class))).thenReturn(lease);

        // When
        Lease returnedLease = leaseService.returnCar(1L);

        // Then
        assertNotNull(returnedLease);
        assertFalse(returnedLease.isActive());
        verify(leaseRepository, times(1)).findById(1L);
        verify(leaseRepository, times(1)).save(returnedLease);
    }

    @Test
    void testReturnCarWhenNotActive() {
        // Given
        Lease lease = LeaseTestDataUtil.getLeaseTestData().get(2);
        lease.setActive(false);
        when(leaseRepository.findById(anyLong())).thenReturn(Optional.of(lease));

        // When & Then
        Exception exception = assertThrows(IllegalStateException.class, () -> {
            leaseService.returnCar(1L);
        });
        assertEquals("Lease is already completed.", exception.getMessage());
    }

    @Test
    void testListActiveLeases() {
        // Given
        List<Lease> activeLeases = LeaseTestDataUtil.getLeaseTestData().stream()
            .filter(Lease::isActive)
            .toList();
        when(leaseRepository.findByIsActive(true)).thenReturn(activeLeases);

        // When
        List<Lease> result = leaseService.listActiveLeases();

        // Then
        assertNotNull(result);
        assertEquals(3, result.size()); // Adjust based on actual data in LeaseTestDataUtil
        verify(leaseRepository, times(1)).findByIsActive(true);
    }

    @Test
    void testDeleteLease() {
        // Given
        Lease lease = LeaseTestDataUtil.getLeaseTestData().get(0);
        when(leaseRepository.findById(anyLong())).thenReturn(Optional.of(lease));

        // When
        leaseService.deleteLease(1L);

        // Then
        verify(leaseRepository, times(1)).delete(lease);
    }

    @Test
    void testDeleteLeaseNotFound() {
        // Given
        when(leaseRepository.findById(anyLong())).thenReturn(Optional.empty());

        // When & Then
        Exception exception = assertThrows(LeaseNotFoundException.class, () -> {
            leaseService.deleteLease(1L);
        });
        assertEquals("Lease not found with ID: 1", exception.getMessage());
    }

    @Test
    void testListLeasesByCustomer() {
        // Given
        List<Lease> leases = LeaseTestDataUtil.getLeaseTestData();
        when(leaseRepository.findByCustomer_CustomerId(anyLong())).thenReturn(leases);

        // When
        List<Lease> result = leaseService.listLeasesByCustomer(1L);

        // Then
        assertNotNull(result);
        assertEquals(4, result.size()); // Adjust based on actual data in LeaseTestDataUtil
        verify(leaseRepository, times(1)).findByCustomer_CustomerId(1L);
    }
}
